rename auth.jsx -> auth/register.jsx
